requirements.txt
pandas==2.0.1
pandasai==0.2.14
python-dotenv==1.0.0
